Expose the episode time axis to interval sources
